+++ r4/src/test/java/io/github/jy95/fds/r4/translators/PeriodPeriodMaxTest.java
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.Locale;
 
 public class PeriodPeriodMaxTest extends AbstractPeriodPeriodMaxTest<FDSConfigR4, Dosage> {
-    
+
     @Override
     protected Dosage generateWithPeriodOnly() {
         Dosage dosage = new Dosage();
